Autenticacion de los endpoints

+++ app_service.py
@@ -7,22 +7,58 @@ class AppService:
     
     # ------------------- TASKS -------------------
 
-    def get_tasks(self):
-        data = self.database.get_tasks()
-        return data
+    def get_tasks(self, user, password):
+        ps = self.database.getUser(user)
+        if len(ps) == 0:
+            return "El usuario no existe"
+        elif ps[0][0] == password:
+            data = self.database.get_tasks()
+            return data
+        else:
+            return "El usuario no tiene permisos."
+
     
-    def get_task_by_id(self, id: int):
-        data = self.database.get_task_by_id(id)
-        return data
+    def get_task_by_id(self, user, password ,id: int):
+        ps = self.database.getUser(user)
+        if len(ps) == 0:
+            return "El usuario no existe"
+        elif ps[0][0] == password:
+            data = self.database.get_task_by_id(id)
+            return data
+        else:
+            return "El usuario no tiene permisos."
+        
     
-    def create_task(self, task: dict):
-        data_inserted = self.database.create_task(task)
-        return data_inserted
+    def create_task(self, user, password ,task: dict):
+        ps = self.database.getUser(user)
+        if len(ps) == 0:
+            return "El usuario no existe"
+        elif ps[0][0] == password:
+            data_inserted = self.database.create_task(task)
+            return data_inserted
+        else:
+            return "El usuario no tiene permisos."
+        
     
-    def update_task(self, id: int, task: dict):
-        data_updated = self.database.update_task(id, task)
-        return data_updated
+    def update_task(self, user, password, id: int, task: dict):
+        ps = self.database.getUser(user)
+        if len(ps) == 0:
+            return "El usuario no existe"
+        elif ps[0][0] == password:
+            data_updated = self.database.update_task(id, task)
+            return data_updated
+        else:
+            return "El usuario no tiene permisos."
+        
     
-    def delete_task(self, id: str):
-        id_deleted = self.database.delete_task(id)
-        return id_deleted
+    def delete_task(self, user, password, id: str):
+        ps = self.database.getUser(user)
+        if len(ps) == 0:
+            return "El usuario no existe"
+        elif ps[0][0] == password:
+            id_deleted = self.database.delete_task(id)
+            return id_deleted
+        else:
+            return "El usuario no tiene permisos."
+        
+
